+++ src/Rojan.Desktop.Domain/Organizations/Permission.cs
@@ -35,4 +35,8 @@ public enum Permission
     SettingsManage,
     OrganizationManage,
     BranchManage,
+    Approve,
+    Reject,
+    Import,
+    ManageUsers,
 }